Bugfix/menu undef val

+++ src/extension.js
@@ -161,10 +161,8 @@ async function showMenu()
 	}
 
 	vscode.window.showQuickPick(dynamicExtMenu, { title: "Global extension menu", matchOnDescription: true, matchOnDetail: true })
-		.then(val => val.foo())
-		.then(undefined, err => {
-			;
-		});
+		.then(val => { if (val) val.foo(); })
+		.then(undefined, err => {});
 }
 
 function Menu_slectSSHfsProfile()
@@ -183,16 +181,16 @@ function Menu_slectSSHfsProfile()
 
 			vscode.window.showQuickPick(profileList, {matchOnDescription: true, matchOnDetail: true})
 				.then(val => {
-					if (!('sshfs' in CTagSSH_Settings.get())) {
-						CTagSSH_Settings.get().sshfs = {};
+					if (val) {
+						if (!('sshfs' in CTagSSH_Settings.get())) {
+							CTagSSH_Settings.get().sshfs = {};
+						}
+						CTagSSH_Settings.get().sshfs.profile = val.label;
+						CTagSSH_Settings.save();
+						connectToSSH();
 					}
-					CTagSSH_Settings.get().sshfs.profile = val.label;
-					CTagSSH_Settings.save();
-					connectToSSH();
 				})
-				.then(undefined, err => {
-					;
-				});
+				.then(undefined, err => {});
 		} else {
 			vscode.window.showErrorMessage("Not found profile of SSH FS extension!");
 			return;
@@ -301,50 +299,50 @@ async function loadRemoteCTags()
 
 				vscode.window.showQuickPick(ctagsFilesList, {title: "CTags: " + conf.ctagsFilesRemotePath, matchOnDescription: true, matchOnDetail: true})
 					.then(async val => {
+						if (val) {
+
+							const localNewCTagsFile = path.join(vscode.workspace.workspaceFolders[0].uri.fsPath, conf.fileCtags);
 						
-						const localNewCTagsFile = path.join(vscode.workspace.workspaceFolders[0].uri.fsPath, conf.fileCtags);
-						
-						// change color
-						updateStatusBar(CTagSSHMode.RemoteDownload);
+							// change color
+							updateStatusBar(CTagSSHMode.RemoteDownload);
 
-						CTagSSH_VF.downloadRemoteFile(
-							pathPosix.join(conf.ctagsFilesRemotePath, val.filename), 
-							vscode.workspace.workspaceFolders[0].uri.fsPath)
-						.then(filename => {
-							try {
-								// remove old ctags file
-								fs.unlinkSync(localNewCTagsFile);
+							CTagSSH_VF.downloadRemoteFile(
+								pathPosix.join(conf.ctagsFilesRemotePath, val.filename),
+								vscode.workspace.workspaceFolders[0].uri.fsPath)
+							.then(filename => {
+								try {
+									// remove old ctags file
+									fs.unlinkSync(localNewCTagsFile);
 
-								// rename new ctags
-								fs.renameSync(filename, localNewCTagsFile);
+									// rename new ctags
+									fs.renameSync(filename, localNewCTagsFile);
 
-								// read new ctags file into memory
-								readCTags(localNewCTagsFile);
-							} catch (err) {
+									// read new ctags file into memory
+									readCTags(localNewCTagsFile);
+								} catch (err) {
+									// revert color back
+									updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
+									return Promise.reject(err);
+								}
+
 								// revert color back
 								updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
-								return Promise.reject(err);
-							}
+							})
+							.then(undefined, err => {
 
-							// revert color back
-							updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
-						})
-						.then(undefined, err => {
-
-							// revert color back
-							updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
-							return Promise.reject(err.message);
-						});
+								// revert color back
+								updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
+								return Promise.reject(err.message);
+							});
+						}
 					})
 					.then(undefined, err => {
-						updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
 						return Promise.reject(err.message);
 					});
 			})
 			.then(undefined, err => {
 				let a = "" + err;
 				console.error(`Can't read remote folder: ` + a);
-				updateStatusBar(CTagSSH_VF.isConnected ? CTagSSHMode.Connected : CTagSSHMode.NotConnected);
 				return Promise.reject(err.message);
 			});
 	}
@@ -487,12 +485,8 @@ function searchTags()
 			});
 
 			vscode.window.showQuickPick(filteredFiles, {title: "Variants", matchOnDescription: true, matchOnDetail: true})
-				.then(val => {
-					navigateToDefinition(val);
-				})
-				.then(undefined, err => {
-					;
-				});
+				.then(val => navigateToDefinition(val))
+				.then(undefined, err => {});
 			break;
 	}
 }
